% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skygrid.R
\name{prepare_skygrid}
\alias{prepare_skygrid}
\title{Create a data.frame for plotting a skygrid model.}
\usage{
prepare_skygrid(
  height,
  pop_size,
  grid_height,
  trajectory = c("median", "mean"),
  ci = c(0.025, 0.975),
  age_of_youngest = 0,
  max_height = c("lower", "upper", "median", "mean"),
  range_time = NULL
)
}
\arguments{
\item{height}{numeric vector; root height parameter.}

\item{pop_size}{data.frame; population size parameters.}

\item{grid_height}{numeric; grid height (i.e. cutoff).}

\item{trajectory}{character string; Trajectory of the population size.
Either mean or median.}

\item{ci}{numeric vector; the default c(0.025, 0.075) corresponds to the
95\% HPD of the population size.}

\item{age_of_youngest}{numeric; sampling time of the youngest taxon.}

\item{max_height}{numeric; Either lower, upper, median or mean.}

\item{range_time}{numeric vector; time range to plot. -1 if automatic.}
}
\value{
A data.frame
}
\description{
Create a data.frame for plotting a skygrid model.
}
