% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{skyplot}
\alias{skyplot}
\title{Plotting a skyline model.}
\usage{
skyplot(
  data,
  y_log = T,
  alpha = 0.5,
  color = NULL,
  fill = NULL,
  color_by = NULL,
  fill_by = NULL,
  group_by = NULL,
  facet_args = list()
)
}
\arguments{
\item{data}{data.frame; data.frame generated by prepare_sky*}

\item{y_log}{logical; if TRUE, y axis is log10 space.}

\item{alpha}{numeric; opacity of ribbon}

\item{color}{character string; color for line.}

\item{fill}{character string; color for ribbons.}

\item{color_by}{character string; name of column for coloring the line.}

\item{fill_by}{character string; name of column for filling the ribbon.}

\item{group_by}{character string; name of column for grouping in facet.}

\item{facet_args}{list; named list of arguments (other than `facets`) passed
to [ggplot2::facet_wrap()]. Sub-plots are superimposed if NULL and group_by
is specified.}
}
\value{
ggplot2 object.
}
\description{
Plotting a skyline model.
}
